+++ xdp/pp_sock.c
@@ -56,7 +56,7 @@ struct config {
     __u32 xdp_flags;
     int ifindex;
     char *ifname;
-    uint32_t iters;   // number of pingpong packet exchanges
+    uint64_t iters;   // number of pingpong packet exchanges
     uint64_t interval;// interval between two pingpong packet exchanges
     __u16 xsk_bind_flags;
     int xsk_if_queue;
